[liftoff] Make tier-up check platform specific
Instead of using a combination of platform-independend operations (load,
sub, store, compare, jump), implement the core tier-up check in
platform-specific code.
This allows to emit smaller code on x64 and ia32 where we can directly
subtract on the memory operand.

Before:
0x2d721ee74706    46  488b8687000000       REX.W movq rax,[rsi+0x87]
0x2d721ee7470d    4d  8b10                 movl rdx,[rax]
0x2d721ee7470f    4f  83ea46               subl rdx,0x46
0x2d721ee74712    52  0f8824000000         js 0x2d721ee7473c  <+0x7c>
0x2d721ee74718    58  8910                 movl [rax],rdx

After:
0x1947e2a49706    46  4c8b9687000000       REX.W movq r10,[rsi+0x87]
0x1947e2a4970d    4d  41832a46             subl [r10],0x46
0x1947e2a49711    51  0f8822000000         js 0x1947e2a49739  <+0x79>

It also simplifies the platform independent code in the LiftoffCompiler
enormously because we do not need to pass in two temporary registers to
any function that might emit a tier-up check.
Instead we use existing scratch registers per platform or push and pop
registers as required.

R=jkummerow@chromium.org

Change-Id: I0b93a340f24b10c3d93e41f17797cd9186300302