[snapshot] Move the SnapshotCreator impl to src/
.. to a new class SnapshotCreatorImpl. Replace the SnapshotCreatorData
class (which is used to hide implementation details from the API
header) with the SnapshotCreatorImpl instance.

We use this code path in important internal spots such as mksnapshot-
it shouldn't be implemented in api.cc.

The code was slightly refactored in the process, e.g. the use of
GlobalHandles and combining the list of (default) contexts and
callbacks.

Bug: v8:13789
Change-Id: I2347d944d7835546dbe03a84d35e46a364d603b1