[sandbox] Sandboxify WasmIndirectFunctionTable::Targets
The |targets| field used to contain a ByteArray containing raw pointers
to code entrypoints, which is not sandbox compatible as an attacker
could change these pointers to jump to arbitrary code. This CL now
introduces a new FixedExternalPointerArray, which uses the external
pointer table mechanism, and changes the |targets| field to contain such
a (sandbox-compatible) array instead.

Bug: chromium:1395058
Change-Id: I280600eed769a5e620912fe85e42f954042b0b16