[regalloc] Limit quadratic merging of SpillRanges
Even after previous speed-ups to `AssignSpillSlots` (see v8:14133),
the merging of all non-empty `SpillRange`s gives still fundamentally
quadratic compile time in the number of `SpillRange`s.

This CL limits that to linear above a certain threshold and
probabilistically limits the downsides of that (see code comment).

For v8:14133, this results in the exact same number of `SpillRange`s
after merging (namely 989, i.e., the quality of the generated code
should be the same for this example) and reduces compile time in the
`V8.TFAssignSpillSlots` phase from ~1600ms to ~255ms (-84%).

Bug: v8:14082, v8:14133
Change-Id: I912f833c65e48ee8727c498fd40507956a68fed9