[ic] Introduce Normal_Fast state for LoadHandler
When the receiver is an object with property dictionary, we save
target name's dictionary entry index in load's inline cache and check
it before dictionary lookup.

We introduce a new state Normal_Fast for this situation. When the
cached index fails the check, we will transit the state back to
Normal.

Change-Id: Ia6684f77aa155060251e5592ac83f3c5607a1daa