[compiler][arm64] Fix select lowering
Word32Select and Word64Select support, for aarch64, was recently
reverted. There were two problems with the handling of selects:
- The inputs to the select could be incorrectly overwritten.
- The incorrect result register could be selected.

Both these bugs appear when we combine an arithmetic operation and a
compare. This patch fixes the bugs, the latter of which can currently
be triggered for fp selects too.

Bug: chromium:1451929
Change-Id: I2daf088f8a892c77659dea390bea8ff00abd4188