[sandbox] Sandboxify WasmIndirectFunctionTable::Targets
The |targets| field used to contain a ByteArray containing raw pointers
to code entrypoints. This is not sandbox compatible as an attacker
could change these pointers to jump to arbitrary code. This CL now
introduces a new ExternalPointerArray type, which uses the external
pointer table mechanism, and changes the |targets| field to contain such
a (sandbox-compatible) array instead.

Most of the code changes in this CL come from the new
ExternalPointerArray which, in contrast to the ByteArrays previously
used, requires special handling by the GC to mark the referenced entries
in the external pointer table as alive.

Bug: chromium:1395058
Change-Id: I280600eed769a5e620912fe85e42f954042b0b16