[atomic,riscv32] fix i64.atomic.rmw32.xxx_u on riscv32.
Liftoff on riscv32 do not clear high reg when handling
i64.atomic.rmw32.xxx_u and thus get wrong result. This
CL add instructions to set high reg of a pair, which
hold an i64 on riscv32, to zero so that unexpected
value in the high bits is cleared.
Besides, riscv64 does not pass the regress test yet, so
it will skip the added test. Further CL will be
committed to solve the underlying bug on riscv64.

Bug: v8:14113
Change-Id: I9d4f40495eebd950dd59d6f3731adf8a042f0fd2