[wasm] Add some more CHECKs for backing stores
See the attached issue; we get a CHECK failure for
`is_wasm_module implies backing_store && backing_store->is_wasm_memory()`.

This CL splits that CHECK in two and adds a few more.

The BackingStore still looks correct in
`GlobalBackingStoreRegistry::UpdateSharedWasmMemoryObjects`, but then
either was replaced by the {EmptyBackingStoreBuffer} or is not marked as
wasm any more a bit later on.

R=ahaas@chromium.org

Bug: chromium:1445003
Change-Id: I2b8d665627b54a867ac55a630fb015832fb13d60