[turbolizer] Turboshaft: Show block and input node label in tooltip
This extends the tooltip to not only list the input ids but also their
origin block and label.
Previously:
  Inputs: 163, 4
Now:
  Inputs: [B27] 163 WordBinop(162, 161), [B0] 4 Constant

This puts more information in the tooltip and makes it easier to find
the input nodes in large graphs due to the block label.
For nodes with many inputs this gets a bit more convoluted, however
the plain index list is already in the main text of the instruction in
the block which is hovered to see the tooltip.

Bug: v8:12783
Change-Id: Id6b942ae0809a9f3e7ef59eb7f48570a10c7b68a