[wasm] Unify signature serialization into a PodArray
Currently there exist two serialization formats of a wasm::FunctionSig
into a PodArray. The one in the WasmCapiFunctionData and the on in
WasmJSFunctionData. This CL replaces both serialization formats with a
new one.

The new format stores the return count in the first slot, followed by
the return types and the parameter types. Since the PodArray is typed
to wasm::ValueType, the return count is embedded into a wasm::ValueType
with wasm::ValueType::FromRawBitField.

R=jkummerow@chromium.org

Bug: v8:14035
Change-Id: I9e5b1bbb4936bb810da4e8a4549eb3fdd0c253f3