[turboshaft] Enable MachineOptimization during MachineLowering
Enabling MachineOptimizationReducer during MachineLoweringReducer
enables LateLoadElimination and StoreStoreElimination to be more
powerful (since they look at the input graph, they see the output
of MachineLoweringReducer, and don't benefit from the
MachineOptimizationReducer in their stack).
However, this also seemingly caused a lot of Chrome crashes, which
were hard to investigate due to the lack of reproducer.

We'll thus now enable this behind an experimental flag, which we'll
fuzz to hopefully figure out what's wrong.

Bug: v8:12783
Change-Id: I00e5416f96a81eba9a33c79c49648d02604801a6