[maglev] Introduce MapCompare::JumpIfNotDeprecated
The only use of `MapCompare::GetMap` method is moved into a new
`MapCompare::JumpIfNotDeprecated` helper. `MapCompare::GetMap` is marked
private and is renamed to `RetrieveMap`.

The goal is to ensure that `MapCompare::JumpIfNotDeprecated` is the
only funciton that calls into (now private) `RetrieveMap`. This allows
for more efficient implementation of `MapCompare`. In particular,
ARM64 version of `MapCompare` now delays decompression until
`RetrieveMap` is called, which then decompresses map pointer in-place.

Bug: v8:7700
Change-Id: I97a9cf2255f817024abf004ff4561e1c48df6f34