[wasm] Limit generic wasm-to-js wrapper to 2 returns
The generic wasm-to-js wrapper writes multi-return values into incorrect
slots on the stack. With Liftoff this does not matter, because Liftoff
access return values sp-relative, but with TurboFan some returns are
incorrect.

Details:
A wasm function which calls another function with multi-return has to
reserve stack space for the return values of the other function. This
stack space overlaps with the stack space used for incoming parameters.
If there are more parameters on the stack than returns, then the returns
don't use all the stack space. The current implementation of the wrapper
assumes that the lower addresses of the stack space are used for returns
then. This assumption is apparently wrong.

R=thibaudm@chromium.org

Bug: v8:14217
Change-Id: I7b98cdd380a588094430d31f130188b006cb285c