Remove AllowLargeObjectAllocation
Excerpted from chromium:1458663#c18 by mlippautz:

AllowLargeObjectAllocation was previously necessary as a callsite from
e.g. CSA would request a new object and perform optimizations based on
the assumption that the object is allocated in the young generation.
Getting an object in LO space in the old generation would violate the
invariants and lead to UAF via missing write barriers. The flag
indicated that the callsite considers this case and can deal with LO
space objects.

Since the introduction of the LO space for the young generation no
invariants are violated anymore. An object in regular new space is
treated the same as an object in LO new space from GC perspective. As
such, the previous issue does not exist any longer.

Fixed: v8:14232
Change-Id: Ia596edeaa094849c8051daabd842881309354a67