[regalloc] Avoid repeated search in ResolveControlFlow
`FindConnectableSubranges` repeatedly called `GetChildCovers` for the
current block. This CL pulls the search in front of the inner loop.
(This also makes `FindConnectableSubranges` a bit pointless, so I
inlined it and added some more comments.)

This reduces profiler samples in `ResolveControlFlow` for the
JetStream2 octane-zlib benchmark from ~62k to ~44k on an x64
workstation and should thus recover most of a slight regression of a
previous cleanup CL (https://crrev.com/c/4701428).

When running all of JetStream2, this CL reduces the total samples in
`ResolveControlFlow` from ~1.49B to ~970k (-35%).

Bug: chromium:1468235, v8:14082
Change-Id: I830bbcdce9d1c6f2fc26064499690c115d185ce4