[turboshaft][js] Add reducer to make int64->int32 truncations explicit
The reducer inserts explicit truncation operations between nodes
expecting an int32 but receiving an int64 input.
Once, the Turboshaft graph is created from source and not from a
Turbofan graph, this should be done directly by the graph builder.

This change does not ensure that following reducers do not insert new
patterns of implicit truncation.

Bug: v8:12783
Change-Id: Iad6f9b1a612c80a9d0722de630b5a7644248f586