[heap] Free LABs on incremental marking start
Free LABs in the safepoint for starting incremental marking. This
avoids having LABs on allocation candidates. While this isn't
necessary for correctness, it avoids corner cases when giving back
unused memory of a LAB to the free list.

It also helps with not creating more additional work for evacuators
due to allocations. So far we did not consider pages with main thread
LABs as evacuation candidates. However, we didn't handle background
threads. Freeing all LABs means that not even background threads can
have LABs on evacuation candidates anymore.

Bug: chromium:1468458
Change-Id: Ib16389157eaf14b1bb7d897242bafa559bbc964b