[wasm][test] Always call through the CEntry stub
The cctests were calling Wasm code directly through a specially compiled
wrapper that was only used in cctests. Also the generated code for traps
was different in that it called a C callback and then returned instead
of unwinding the stack via a runtime function.

This CL removes this test-only code and always calls through the CEntry
stub such that stack unwinding works as in production.

R=ahaas@chromium.org

Bug: v8:14235
Cq-Include-Trybots: luci.v8.try:v8_linux64_ubsan_rel
Change-Id: I446cdcc39e3132fd9f662453332e31ec2433b0e8