[tagged-ptr] Remove (heap-)object include from tagged.h
Make tagged.h work with forward declarations of Object/HeapObject, by
moving the methods that need the definition to be immediately after that
definition.

This allows us to use Tagged<Object> without a dependency on the
objects.h header, as demonstrated in identity-map.h here. It will also
allow Object/Smi/HeapObject to return Tagged<*> in their methods.

Bug: v8:12710
Change-Id: Ic28cdcf5b0e047b8f0296c3c47d497efb85ae771