[wasm] Generate better link error messages
Always include the import index, module name, and import name, if
present.
In return, drop a redundant "error: " part of the message.

This makes the code easier to navigate by removing two macros that
define four new methods. Instead we just call one of two new methods to
generate a consistent name for identifying an import.

R=ahaas@chromium.org

Change-Id: Id8653adf2d1e0dd3d5bae11970b7e99daec535b5