[maglev] Fix dcheck in for-reach map reset
In for-each we tried to reset the node type to the state before the
loop using the SetPossibleMaps api. We can end up setting a node type
that is stricter than what can be inferred from the possible maps,
simply because we might have more information in the node type from
other sources than just map checks.

Overal the solution is to simply not try to do anything to the node type
as it is monotonically narrowing and thus does not need resetting.

Bug: chromium:1470292
Change-Id: Ica14a89cfd3b297efea6176717e00c9c1b745944