[shared-struct] Make shared object constructors have instance prototypes
For POJOs, having a null prototype requires marking the constructor as
having a non-instance prototype. For shared objects, this is not
necessary as all shared objects have an instance map (even though the
prototype is null).

This also fixes a bug where currently shared object constructors are
incorrectly mutating the strict_function_with_readonly_prototype_map.

Bug: v8:12547
Change-Id: I04729b87cd887193941ebcc490de62a79c1a545c