[deopt] Fix MaglevFrame::GetBytecodeOffsetForOSR
The problem here was that in the case of builtin continuation frames
the osr offset in the DeoptimizationData contains a builtin id and not
an actual bytecode offset. To be sure we hit the correct frame, we
actually must parse the frame translation metadata and find the
innermost interpreter frame.

Drive-By: Move a bunch of logic for parsing FrameTranslations to its
iterator so we can use it in the fix.

Bug: chromium:1471503, v8:14273
Change-Id: I4837785c8bb479d66b6e07731f1810c3f91b8812