[wasm] Clean up, document, and fix handling of dying isolates
The actual change of this CL is the removal of logic to remove code
objects that are present in the {potentially_dead_code} set from the
current GC's {dead_code}.
This code is there since a long time (https://crrev.com/c/1559738), but
it looks wrong: Only the isolate is dying, not the {NativeModule}, and
by removing it from the GC's {dead_code} list we are preventing the code
from being garbage-collected in the current GC.

Instead we should do the equivalent of the {WasmGCForegroundTask}: The
current isolate is not using any code any more (equivalent of an empty
stack), so all potentially dead code is actually dead from this
isolate's POV. So we just remove it from the current GC without
reporting any live code.

All other changes are just renamings, reorderings, and adding comments.

R=thibaudm@chromium.org

Change-Id: Ibfc542b24d057e5b086d166476c10f43007eaead