[utils] Add base::SmallMap data structure
We sometimes use many short-lived and small maps during
compilation, e.g., in the register allocator. One can improve both
runtime and memory usage over a std::(unordered_)map by using linear
search in an inline array when there are less than a small number of
elements, and fall back to a "proper" map once it grows beyond that
threshold.

Chromium already has such a datastructure, so this is simply copied from
https://source.chromium.org/chromium/chromium/src/+/main:base/containers/small_map.h
and
https://source.chromium.org/chromium/chromium/src/+/main:base/containers/small_map_unittest.cc

The only modifications are to align with the V8 codebase and fix linter

warnings:
- put in `v8::base` namespace.
- rename `small_map` to `SmallMap` to fit with the naming of our other
collections (similarly the other classes in `small-map.h`).
- use `V8_INLINE` instead of `inline`.
- annotate copy ctor/assignment operator with `V8_NOEXCEPT`.
- replace `raw_ptr` in iterator with raw pointer.
- make global constant `kUsingFullMapSentinel` a static constant member
to not pollute the global namespace.
- run `git cl format`.
Change-Id: Ic918e80bf1f020a38c07bdbefb087234a7a9b499