[utils] Add SmallMap::try_emplace
`try_emplace` is useful over `emplace` since it guarantees nothing is
done if the key is already present in the map. In particular it doesn't
construct a `value_type` just for lookup. This also brings `SmallMap`'s
API in line with std::(unordered_)map which got this with C++17.

A follow-up CL in the register allocator will use this. There, it made
a measurable difference since the value type is large.

Change-Id: I73212ec8f73e9c9143c9e265104b93707c724d60