[regalloc][cleanup] Use SmallZoneMap instead of custom class
`RangeRegisterSmallMap` was essentially already a `SmallMap` (i.e.,
storing key-value-pairs in an unsorted inline array because it is
expected to be of small size).

This CL thus removes redundant code. It has no locally measureable
performance impact on JetStream2 (as expected).

Drive-by: Also add some DCHECKs for better documenting the invariant
that a `TopLevelLiveRange` is inserted at most once in `to_be_live`.

Bug: v8:14082
Change-Id: I016b480281c404c7dd36ab6837a492840a0b98aa