[heap] Use stack marker for conservative stack scanning
This CL refactors conservative stack scanning so that it always scans
from the stack marker to the stack start. It assumes that a stack marker
has already been set, when iterating through stack roots. The method
Stack::SetMarkerIfNeededAndCallback must be used to set the marker: it
only does so if the marker has not already been set. The callback is
required because the stack marker is set inside the trampoline for
saving callee-saved registers on the stack.

Using the stack marker instead of scanning the entire stack is useful
in cases where it is important to consistently scan the same part of
the stack successively. This CL addresses three such cases:

1. During GC, for marking and marking verification.
2. During snapshot creation, GC and read-only promotion should scan the
   same part of the stack. The same during read-only promotion, when
   determining the promotees and updating the pointers.
3. In the heap profiler, during snapshot generation, the GC and the
   filling of references should scan the same part of the stack.
   The same when querying for objects.

The CL also cleans up the trampoline-related logic that was implemented
in the local heap, as well as the now redundant scan stack modes.

Bug: v8:13257
Change-Id: I36e68e4d5201def75ae9841dfcd9c0fd8fdf329e