[sandbox] Add a TRUSTED_SPACE and TRUSTED_LO_SPACE to the V8 heap
The trusted spaces will contain objects that are trusted by (parts of)
V8 to not have been manipulated by an attacker. Examples of such objects
include BytecodeArrays or DeoptimizationData, both of which may cause
stack corruption if modified, thereby allowing an attacker to break out
of the sandbox.

In order to defend against this, the trusted spaces can later be moved
out of the sandbox and into its own pointer compression cage. References
from inside the sandbox to trusted objects will use a pointer table
indirection (an "indirect pointer"), which ensures memory-safe access to
them. Apart from that, the two trusted spaces behave like the old space
and lo space, respectively.

Bug: chromium:1473677
Change-Id: I4eb735359c9b71c86f2cbb9bc7da4b89b58e9f5b