[sandbox] Introduce the IndirectPointerTable (IPT)
This pointer table will be used to reference V8 HeapObjects located
outside of the sandbox in a memory-safe way, similar to how the
ExternalPointerTable (EPT) is used for external (non-V8) objects.

The IPT is also essentially a generalization of the CodePointerTable
(CPT), which is already used to reference Code objects. An IPT can then
in principle be used to reference any kind of HeapObject.

This CL does not yet use the IPT. This will happen in follow-up CLs.

Bug: chromium:1473677
Change-Id: Ia50ec93297f06a24d176ed4beb17d89f21337d26