[parser] Inherit language mode for eval'd scripts
When constructing UnoptimizedCompileFlags from an exsting eval Script,
the initial language mode should be inherited from the SFI that called
eval.

This fixes a debugger crash where the scope iterator gets confused
when it expects a function scope to require a context with none
present as the function got wrongly reparsed in sloppy mode instead
of strict mode.

R=leszeks@chromium.org

Bug: chromium:1474272
Change-Id: I176c01f0299c5d0ab24a4440f950a1c95d08a91a