[sandbox] CHECK that EPT entries have the expected tag in ASan builds
The external pointer table mechanism relies on the fact that a tag
mismatch results in a pointer pointing to a non-canonical address, and
that this always results in a (harmless) crash when dereferenced. The
sandbox crash filter than filters out crashes at non-canonical addresses
as these are considered "safe". However, when ASan is active, it can
happen that ASan's instrumentation itself crashes with a segmentation
fault when reading from the shadow memory for a memory access. This
will, in turn, cause a crash at a _canonical_ address, thus producing a
false-positive report.

To ignore these false-positives, this CL now adds a CHECK into the
ExternalPointerTable::Get routine when ASan is active. This check will
fail if it detects a tag mismatch (which would otherwise produce an
invalid pointer). In non-ASan builds, this check is a DCHECK instead.

Bug: chromium:1475831
Change-Id: I58270a777e2e1a34bc3d8ee48e1e7c3209f9f507