[shadow stack] lazy deopt changes
This CL is a rebase of Michael Stanton's original CL
https://chromium-review.googlesource.com/c/v8/v8/+/3295457.
I'm uploading it as part of my CL series that prototype GCS
support for arm64.

For lazy deoptimization, we currently patch any active addresses on
the stack to jump to their deoptimization entries. This doesn't work
with a shadow stack. Therefore, additional codegen will check for
deoptimization after returning from calls that have lazy deopt
state.

Bug: v8:11246
Change-Id: If9dce2baaa5f951ae894890cada651e7d15c1123