[heap] Fix LargeObjectSpaceObjectIterator
This CL adds a DCHECK to ensure that the LargeObjectSpaceObjectIterator
never returns filler of free space objects, just like all other space
object iterators. This is based on the assumption that, when we are at
a safepoint, large pages cannot contain fillers or free space objects.
The CL refactors some tests and related machinery for stressing
incremental marking and concurrent allocation, to honor this assumption.

Change-Id: If92c514797ac96d4292cfcc6f804f47357552519