[heap] Avoid scheduling a minor gc task during loading.
The task would be scheduled, and when it ran if loading was in progress,
it would reschedule itself and bail out. The rationale was that the
rescheduled task would run at a later time and loading may be done by
then. However, it appears the tasks are just executed one after the
other repeatedly.

Bug: v8:12612, chromium:1477834
Change-Id: I3a6c93d466258f60cbe03e4e78d9203f71be2de8