[sandbox] Directly associate pointer tags with slots
We now store the ExternalPointerTag/IndirectPointerTag inside the
respective slot class instead of having to pass it around separately.
As such, code that was previously doing something like

    void VisitExternalPointer(Tagged<HeapObject> host,
                              ExternalPointerSlot slot,
                              ExternalPointerTag tag) {
      ... = slot.load(tag);
    }

now simply becomes:

    void VisitExternalPointer(Tagged<HeapObject> host,
                              ExternalPointerSlot slot) {
      ... = slot.load();
    }

as the slot class stores the tag associated with that slot.

This is probably more natural since every slot must have an associated
tag and so can only ever contain pointers with that tag. Or in other
words, we can never access the same {external|indirect} pointer slot
with two different tags.

Bug: chromium:1473677
Change-Id: I378cc2c8b502dbaa141945a1b32c09d8920da39f