[sandbox] Add a dedicated PageAllocator for the trusted spaces
This CL adds the trusted_page_allocator() to the MemoryAllocator and
uses it for pages that belong to either TRUSTED_SPACE or
TRUSTED_LO_SPACE. Currently, that page allocator is the same as the
default page allocator (i.e. allocates pages inside the sandbox), but in
a follow-up CL it will become a separate page allocator that allocates
its pages outside of the sandbox.

Bug: chromium:1473677
Change-Id: I0440f060a0fe94cb917d6eb9c811705f0fa02895