[turboshaft] Handle immutable loads/store in load-elimination
Immutable loads and stores should never be invalidate in load
elimination, as the loaded value never changes. With this CL
invalidation is avoided for immutable loads and stores by not calling
`AddKeyInBaseOffsetMaps` for these loads and stores. This is achieved by
using `SnapshotTable::Set` instead of `ChangeTrackingSnapshotTable::Set`
to add the immutable loads and stores to the `MemoryContentTable`.

R=dmercadier@chromium.org

Bug: v8:14108
Change-Id: I0ee6e6861cc3a4bb7124da3727b4c1e5d6a510ed