[heap] Set MemoryBalancer limit in Heap::RecomputeLimits
So far RecomputeLimits was setting heap limits using the default
heuristics even with --memory-balancer. This CL refactors
memory-balancer.cc such that it sets the limit in
Heap::RecomputeLimits.

In order to do see GC speed and allocation rate need to be updated
before recomputing the limit at the end of a full GC. We also now
track the start and end of the atomic pause in GCTracer.

Bug: v8:13842
Change-Id: I042cda9e87901c89b7cd83432eef53b50cfb1838