[tagged-ptr] Make Smi AllStatic
Start the port of objects to C++ layout with Smi, which is not an object
with layout at all but rather a tag class indicating that the Tagged<>
represents a tagged integer and not a tagged pointer.

This means that Smi becomes an all-static class of helper methods, and
is otherwise just used as a tag. Since it no longer inherits from
Object, we have to turn various std::base_of checks into our own
template specialized "is_subtype" trait -- this has been promoted
from tnode.h to be a generic V8 type trait, and slightly modernised.

Also, the same thing for TaggedIndex.

Bug: v8:7700
Change-Id: Iac338c78beb14f887611e853cf3ba4a3dc41312c