[codegen][x64] Stress smis by clobbering unused bits
In pointer compressed builds the upper 32 bits of smis do not have a
defined value.

We would like to trigger type confusion bugs between smi and int32
earlier. To that end this change introduces a clobbering scheme where
the unused bits are xord with a recognizable bitmap (0x515151). E.g.,
this will cause comparison between Smi::zero() and 0L to return false.

Unfortunately we cannot apply the clobbering to smi constants, because
currently the codegen sometimes relies on their 64 bit value, e.g.
when using them for builtin offsets.

Bug: v8:7700
Change-Id: Ic3d57db8b83671bc1ee6aec5fc89a9d811624f62