[sandbox] Introduce IndirectlyReferenceableObject type
This type describes objects that can be referenced through indirect
pointers. Code object, which are still the only objects that can be
referenced in that way, now inherit from this type.

With indirect pointers, an object is referenced through an entry in a
pointer table, and the referring object only stores an index into the
table. Since we want to have one table entry per referenced object,
*not* one table entry per reference, there must be a way to obtain the
existing table entry from the referenced object. With the
IndirectlyReferenceableObject this now works through the table entry
index stored at offset 4 in every such object. As such, the routines to
write into an indirect pointer slot simply become:

void WriteIndirectPointerField(int offset,
                               IndirectlyReferenceableObject value) {
    auto handle = value.ReadField<uint32_t>(4);
    WriteField<uint32_t>(offset, handle);
}

Bug: chromium:1473677
Change-Id: I2cdfe397b869cc496a1dc94ea51fa01a37a8d4d8