[heap] Move slow path allocation functions into cc file
This CL moves the slow allocation methods in SpaceWithLinearArea
out of the header into the .cc file and makes them private.

For tests a new method AllocateRawForceAlignmentForTesting is added,
which allows to test aligned allocations and for this purpose also
forces the alignment code path even when it's disabled at build
time.

Drive-by: Move MemoryChunk::FreeListsLength into memory-chunk.cc.

Bug: chromium:1480975
Change-Id: I42f735b16be312ba365ba1f226004601fb1fbaac