[sandbox] Introduce the TrustedByteArray
A TrustedByteArray is a trusted object that simply contains a fixed
number of raw bytes in its body. It uses a new instance type and a
custom Map, which is, for now, located inside the sandbox in the regular
heap. However, in the future, we might want to move the Map into trusted
space as well.

Currently, TrustedByteArrays are exposed to untrusted objects inside the
sandbox (so they extend ExposedTrustedObjects). This will make the
transition to trusted space easier, as object graphs can be migrated
incrementally. Eventually, we might want to make TrustedByteArrays
"internal" trusted objects that can only be referenced through some
other, exposed trusted object.

Bug: chromium:1473677
Change-Id: Ifcaed709b5448a4d9b0cae1772146866241a71db