[json] Optimize whitespace check in JSON.parse
The central loop in the JSON parser starts by calling SkipWhitespace,
which must check whether the current character is whitespace and loop
until a non-whitespace character is found. This implementation makes
sense if whitespace is common, but I expect that most websites strive to
avoid unnecessary whitespace. In this change, I propose moving the
whitespace-eating loop into a case of the `switch` statement, to reduce
the number of instructions between the loop top and the `switch`. I
measured an improvement of about 5% on the parseLargeString case in
v8:8748.

Bug: v8:8748
Change-Id: I194d5da90caaefe8b4df6f341a55432efef46b0f