[wasm-gc] Final binary encoding
This patch brings the official final binary encoding for WasmGC
instructions and types to V8.
This is a breaking change. Modules compiled for the "prototype
spec" (bit.ly/3cWcm6Q) will no longer work.
Unfortunately this also needs to shift the encodings for the
"stringref" and "stringview_wtf8" types, to resolve collisions.

Bug: v8:7748, v8:12868
Change-Id: I6d81c34867066f0388f4f2e44ec521d7a594d203