[json] Generate consistent error messages for expected ':'
There are two places in the JSON parser that expect the next character
to be ':', since the first property in an object is handled in a
different location than all of the other properties. I noticed that
those two places generate different error message, so this change makes
them match.

Change-Id: Iae34e018e6ce38124154cd32ac42a36ba04ac0bd