[maglev] Fix slack tracking race
This is a port of 7effdbf988a7d18f4cbe9dece94929ff76eae600
(https://crrev.com/c/4640063), which fixed the same bug for Turbofan.
See the description of this other commit for details about the bug.

In a nutshell: if slack tracking finishes during compilation, we could
fail to realize that a backing store should be extended. The fix:
detect such cases and bail out.

Fixed:1482079
Change-Id: I2b93f1e7e75aeb0e15fba949a431815066dcbcd1