[heap] Refactor StrongRoot(Block)Allocator
This CL cleans up the implementation of StrongRootBlockAllocator,
removing redundant or deprecated things and aligning with C++ >= 11.
The allocator is now renamed as StrongRootAllocator, has become a
template and has StrongRootAllocator<Address> as a special case,
registering strong roots. All other instances simply behave as
std::allocator.

The CL also adds tests to verify that the allocator behaves as
intended, both stand-alone and when used with standard C++ containers.

Change-Id: I8f16a6022918db225c097a8fa742762101aee146