[tagged-ptr] Update debug_helper to use TaggedMember
Currently, cctest/test-v8windbg is failing on builds without pointer
compression. This is because it tries to interpret the in-object
properties for a JSObject as an array of Object, but Object is now
AllStatic.

To fix the problem, this CL updates the way that fields are represented
in debug-helper.h. Previously, there was both a type and a
decompressed_type for each field, to represent both the actual runtime
type of the field and the type of the heap object it points to, if
relevant. However, there's now a much cleaner way to represent the type
of such fields: TaggedMember<T>. Using that template type, we needn't
include a separate decompressed_type, and we can make the behavior of
debug_helper more consistent between builds with and without pointer
compression.

There are still some other open issues in v8windbg such as its failure
to understand Tagged<T> objects, but this change is just the minimum to
get the test passing again.

Bug: v8:12710
Change-Id: I998b88c48cf3a55f8450c1ef1afb0774772649aa