[turboshaft] Lazily increment use_counts
With this CL, use_counts are only incremented once
required_when_unused operations are emitted (at which point we
recursively increment the use_count of any of the input of the
operation whose use_count was 0).

This will allow for better cheap dead code elimination when there is a
chain of unused operations.

Bug: v8:12783
Change-Id: I2bced05687de039c231feee390a8105f6b169bb7