Don't check for implicit fallthrough in jump threading
The implementation of jump threading includes some logic to check
whether a block falls through into the following block. However, this
logic is unnecessary, because earlier phases of the compiler never
construct blocks with implicit fallthrough. Blocks with a single
successor, even if the successor is the next block in order, end with
the kArchJmp instruction. Furthermore, the fallthrough logic in
JumpThreading is incorrect: ComputeForwarding treats implicit
fallthrough as a jump to the next block in reverse post-order, whereas
ApplyForwarding treats implicit fallthrough as a jump to the next block
in assembly order. (I am entirely to blame for this inconsistency; see
https://crrev.com/c/3039261 .)

In this change, I propose that we remove the code which checks for
fallthrough, since it is both unnecessary and incorrect.

Change-Id: I36410ebd56e7cf0f044f2a5e53f8e35f57d9ea3e