[heap] Fix failure in EvacuationVerifier
Finishing sweeping invokes the EvacuationVerifier with --verify-heap.
This verifier needs all spaces and pages to be iterable, however this
is not the case with active LABs.

This CL switches the order of operations in Heap::MakeHeapIterable(),
such that LABs are made iterable and only finishing sweeping
afterwards. That way the full heap should be iterable.

Bug: chromium:1480975, chromium:1488263
Change-Id: Id884fa2eef1f858538e4a54a168556d80f0c6abf