[turboshaft][wasm] Implement inlining infrastructure
This CL implements inlining for turboshaft wasm. Inlining is
implemented as part of graph generation.
We add a `Mode` to the graph interface to signify whether a function
is not inlined, inlined with an exception handler, or inlined without
one. Inlined functions are passed parameters from the caller and jump
to a special return block instead of returning. Also, thrown exceptions
are passed to the caller, and handled there like function call
exceptions.
Node origins during compilations are updated to encode an inlining id
in addition to the function offset.
This CL does not handle:
- proper inlining heuristics; instead the first 2 function calls are
  inlined,
- inlining tail calls (although it does handle tail calls in the
  inlinee),
- speculative inlining.

Bug: v8:14108
Change-Id: Ia4be29b7ff76c9b2bd08964f76d28743a38d85cd